Extract 3D browse mode into html-generator-3d.ts
html-generator.ts had grown to ~2.7k lines, mostly the embedded 3D
browse-mode assets. Move them verbatim into a dedicated module:

- asd3dStyles  - the 3D CSS block (static)
- asd3dScript  - the 3D browse-mode IIFE (static, ~1.7k lines)
- asd3dOverlay(safeAlpsTitle) - the modal markup (only the title is interpolated)

generateHtml() now just splices these in. The generated HTML is
byte-for-byte identical to before (verified by diff); html-generator.ts
drops from 2764 to 954 lines. Build clean, all 127 tests pass.
